+++ packages/perseus/src/widget-ai-utils/label-image/label-image-ai-utils.test.ts
@@ -3,6 +3,8 @@ import {userEvent as userEventLib} from "@testing-library/user-event";
 
 import {renderQuestion} from "../../widgets/__testutils__/renderQuestion";
 
+import {getPromptJSON} from "./label-image-ai-utils";
+
 import type {PerseusRenderer} from "../../perseus-types";
 import type {UserEvent} from "@testing-library/user-event";
 
@@ -68,7 +70,7 @@ const textQuestion: PerseusRenderer = {
     },
 };
 
-describe("label image widget", () => {
+describe("LabelImage AI utils", () => {
     let userEvent: UserEvent;
     beforeEach(() => {
         userEvent = userEventLib.setup({
@@ -76,6 +78,93 @@ describe("label image widget", () => {
         });
     });
 
+    it("it returns JSON with the expected format and fields", () => {
+        const renderProps: any = {
+            markers: [
+                {
+                    answers: ["SUVs"],
+                    label: "The fourth unlabeled bar line.",
+                    x: 25,
+                    y: 17.7,
+                },
+                {
+                    answers: ["Trucks"],
+                    label: "The third unlabeled bar line.",
+                    x: 25,
+                    y: 35.3,
+                },
+                {
+                    answers: ["Cars"],
+                    label: "The second unlabeled bar line.",
+                    x: 25,
+                    y: 53,
+                },
+                {
+                    answers: ["Vans"],
+                    label: "The first unlabeled bar line.",
+                    x: 25,
+                    y: 70.3,
+                },
+            ],
+            choices: ["Trucks", "Vans", "Cars", "SUVs"],
+            imageUrl:
+                "web+graphie://ka-perseus-graphie.s3.amazonaws.com/56c60c72e96cd353e4a8b5434506cd3a21e717af",
+            imageAlt:
+                "A bar graph with four bar lines that shows the horizontal axis labeled Number in the parking lot and the vertical axis labeled Vehicle Type. The horizontal axis is labeled, from left to right: 0, 10, 20, 30, 40, and 50. The vertical axis has, from the bottom to the top, four unlabeled bar lines as follows: the first unlabeled bar line extends to the middle of 0 and 10, the second unlabeled bar line extends to 40, the third unlabeled bar line extends to the middle of 20 and 30, and fourth unlabeled bar line extends to 10.",
+        };
+
+        const userInput: any = {
+            markers: [
+                {
+                    label: "The fourth unlabeled bar line.",
+                    selected: ["Vans"],
+                },
+                {
+                    label: "The third unlabeled bar line.",
+                    selected: ["SUVs"],
+                },
+            ],
+        };
+
+        const resultJSON = getPromptJSON(renderProps, userInput);
+
+        expect(resultJSON).toEqual({
+            type: "label-image",
+            options: {
+                markers: [
+                    {
+                        label: "The fourth unlabeled bar line.",
+                    },
+                    {
+                        label: "The third unlabeled bar line.",
+                    },
+                    {
+                        label: "The second unlabeled bar line.",
+                    },
+                    {
+                        label: "The first unlabeled bar line.",
+                    },
+                ],
+                choices: ["Trucks", "Vans", "Cars", "SUVs"],
+                imageUrl:
+                    "web+graphie://ka-perseus-graphie.s3.amazonaws.com/56c60c72e96cd353e4a8b5434506cd3a21e717af",
+                imageAlt:
+                    "A bar graph with four bar lines that shows the horizontal axis labeled Number in the parking lot and the vertical axis labeled Vehicle Type. The horizontal axis is labeled, from left to right: 0, 10, 20, 30, 40, and 50. The vertical axis has, from the bottom to the top, four unlabeled bar lines as follows: the first unlabeled bar line extends to the middle of 0 and 10, the second unlabeled bar line extends to 40, the third unlabeled bar line extends to the middle of 20 and 30, and fourth unlabeled bar line extends to 10.",
+            },
+            userInput: {
+                markers: [
+                    {
+                        label: "The fourth unlabeled bar line.",
+                        selected: ["Vans"],
+                    },
+                    {
+                        label: "The third unlabeled bar line.",
+                        selected: ["SUVs"],
+                    },
+                ],
+            },
+        });
+    });
     it("should get prompt json which matches the state of the UI", async () => {
         // Arrange
         const {renderer} = renderQuestion(textQuestion);
